requirements.txt
google-genai>=1.0.0
groq>=0.9.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
tavily-python>=0.3.0
pdfplumber>=0.11.0
python-docx>=1.1.0
requests>=2.32.0
streamlit>=1.40.0
numpy>=1.26.0
python-dotenv>=1.0.0